This is my third commit to this project.

+++ migrations/sqls/20220809061755-orders-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
-    user_id bigint NOT NULL,
+    user_id INTEGER NOT NULL,
     FOREIGN KEY (user_id) REFERENCES  users(id) ON DELETE CASCADE,
     status VARCHAR(15)
 );
